refactor: extract delivery verification workflow into dedicated service
